+++ pom.xml
@@ -37,8 +37,8 @@
     </sonar.exclusions>
 
     <folio-spring-support.version>10.0.0</folio-spring-support.version>
-    <folio-service-tools.version>6.0.0-SNAPSHOT</folio-service-tools.version>
-    <folio-s3-client.version>3.0.0-SNAPSHOT</folio-s3-client.version>
+    <folio-service-tools.version>6.0.0</folio-service-tools.version>
+    <folio-s3-client.version>3.0.0</folio-s3-client.version>
     <folio-isbn-utils.version>1.9.0</folio-isbn-utils.version>
     <folio-cql2pgjson.version>36.0.0</folio-cql2pgjson.version>
     <opensearch.version>3.6.0</opensearch.version>
