+++ src/main/java/org/folio/search/utils/KafkaConstants.java
@@ -14,5 +14,5 @@ public class KafkaConstants {
   public static final String INSTANCE_SHARING_COMPLETE_LISTENER_ID = "mod-search-instance-sharing-complete-listener";
   public static final String REINDEX_RANGE_INDEX_LISTENER_ID = "mod-search-reindex-index-listener";
   public static final String REINDEX_RECORDS_LISTENER_ID = "mod-search-reindex-records-listener";
-  public static final String REINDEX_RECORDS_FILE_READY_ID = "mod-search-reindex-reindex-file-ready-listener";
+  public static final String REINDEX_RECORDS_FILE_READY_ID = "mod-search-reindex-file-ready-listener";
 }
