+++ src/main/resources/application-dev.yml
@@ -92,6 +92,7 @@ folio:
     upload-range-size: ${REINDEX_UPLOAD_RANGE_SIZE:1000}
     upload-range-level: ${REINDEX_UPLOAD_RANGE_LEVEL:3}
     merge-range-size: ${REINDEX_MERGE_RANGE_SIZE:500}
+    merge-export-batch-size: ${REINDEX_MERGE_EXPORT_BATCH_SIZE:500}
     merge-range-publisher-core-pool-size: ${REINDEX_MERGE_RANGE_PUBLISHER_CORE_POOL_SIZE:3}
     merge-range-publisher-max-pool-size: ${REINDEX_MERGE_RANGE_PUBLISHER_MAX_POOL_SIZE:6}
     merge-range-publisher-retry-interval-ms: ${REINDEX_MERGE_RANGE_PUBLISHER_RETRY_INTERVAL_MS:1000}
@@ -165,7 +166,7 @@ folio:
         group-id: ${folio.environment}-mod-search-reindex-records-group
       reindex-file-ready:
         concurrency: ${KAFKA_REINDEX_FILE_READY_CONCURRENCY:2}
-        topic-pattern: (${folio.environment}\.)(.*\.)inventory\.reindex.file-ready
+        topic-pattern: (${folio.environment}\.)(.*\.)inventory\.reindex\.file-ready
         group-id: ${folio.environment}-mod-search-reindex-file-ready-group
       index-instance:
         concurrency: ${KAFKA_INDEX_INSTANCE_CONCURRENCY:2}
